Improve nav item extensibility
